refactor: PopupMainInformation에 id 필드 추가 및 파일 분리
