+++ Popcorn-iOS/Source/Domain/Entities/MainScene/MainScene/PopupDetailScene/PopupMainInformation.swift
@@ -1,0 +1,17 @@
+//
+//  PopupMainInformation.swift
+//  Popcorn-iOS
+//
+//  Created by 제민우 on 2/19/25.
+//
+
+import Foundation
+
+struct PopupMainInformation {
+    let popupId: Int
+    let popupTitle: String
+    let startDate: Date
+    let endDate: Date
+    let isUserPick: Bool
+    let hashTags: [String]?
+}
